+++ include/pfnet/pfnet.h
@@ -43,11 +43,12 @@
 #include <pfnet/constr_NBOUND.h>
 #include <pfnet/constr_PAR_GEN_P.h>
 #include <pfnet/constr_PAR_GEN_Q.h>
+#include <pfnet/constr_GEN_RAMP.h>
 #include <pfnet/constr_REG_GEN.h>
 #include <pfnet/constr_REG_TRAN.h>
 #include <pfnet/constr_REG_SHUNT.h>
 #include <pfnet/constr_DC_FLOW_LIM.h>
 #include <pfnet/constr_AC_FLOW_LIM.h>
-#include <pfnet/constr_GEN_RAMP.h>
+#include <pfnet/constr_BAT_DYN.h>
 
 #endif